Explicit cast when snprintf'ing an uint64.  Prevents warnings on platforms where int64 is long not long long.  ok djm@

+++ usr.bin/ssh/auth2.c
@@ -1,4 +1,4 @@
-/* $OpenBSD: auth2.c,v 1.145 2018/03/03 03:15:51 djm Exp $ */
+/* $OpenBSD: auth2.c,v 1.146 2018/04/13 03:57:26 dtucker Exp $ */
 /*
  * Copyright (c) 2000 Markus Friedl.  All rights reserved.
  *
@@ -210,7 +210,8 @@ user_specific_delay(const char *user)
 	u_char *hash = xmalloc(len);
 	double delay;
 
-	(void)snprintf(b, sizeof b, "%llu%s", options.timing_secret, user);
+	(void)snprintf(b, sizeof b, "%llu%s",
+	     (unsigned long long)options.timing_secret, user);
 	if (ssh_digest_memory(SSH_DIGEST_SHA512, b, strlen(b), hash, len) != 0)
 		fatal("%s: ssh_digest_memory", __func__);
 	/* 0-4.2 ms of delay */
